[refactor] #1 ControlMenu 컴포넌트 분리
