+++ src/components/ControlMenu.js
@@ -1,0 +1,20 @@
+import React from "react";
+
+const ControlMenu = React.memo(({ value, onChange, optionList }) => {
+  return (
+    <select
+      className="ControlMenu"
+      value={value}
+      name={value}
+      onChange={(e) => onChange(e.target.value)}
+    >
+      {optionList.map((it, idx) => (
+        <option key={idx} value={it.value}>
+          {it.name}
+        </option>
+      ))}
+    </select>
+  );
+});
+
+export default ControlMenu;
